added in functionality to map view
added in markers and popups to those markers in the quest map view. When the user first opens up the app, their location gets set as a purple marker on the map and the map is also centered around their location. Need to work on what to do when 1. Location changes, and 2. how to get the map to center / zoom in a way that all quests are visible (this should change dynamically based on the quests available).